change to Utils instead of Registry, add hacks for creating custom rule types
